// src/app/models/borne.model.ts
export interface Borne {

  id: number;

  uuid: string;

  title: string;

  address: string;

  city: string;

  province: string;

  postcode: string;

  latitude: number;

  longitude: number;

  operator: string;

  operatorWebsite: string;

  status: string;

  isOperational: boolean;

  usageCost: string;

}